requirements.txt
alabaster==0.7.9
Babel==2.3.4
cycler==0.10.0
docutils==0.13.1
imagesize==0.7.1
Jinja2==2.8
MarkupSafe==0.23
matplotlib==1.5.3
numpy==1.11.3
olefile==0.44
Pillow==4.0.0
py==1.4.32
Pygments==2.1.3
pyparsing==2.1.10
PyQt5==5.7
pyssim==0.3
pytest==3.0.5
python-dateutil==2.6.0
pytz==2016.10
requests==2.12.4
scipy==0.18.1
sip==4.18.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.1
sphinx-rtd-theme==0.1.9
